Smarter Reward Delivery and ActiveStop Updates
Now, reward delivery uses toggles which activate if an active lick or stop occurred in the reward zone. Before, a lick had to occur on the frame when reward was delivered, and a lick had to be performed after stop was completed.

Also - propagated updates to activestopping through vrControlRunExperiment and other downstream code.